requirements.txt
hyperliquid-python-sdk
python-telegram-bot==21.5
loguru==0.7.2